changed to full paper support and paper loader
