--- bukkit/src/main/resources/plugin.yml
@@ -1,7 +1,0 @@
-name: ${project.name}
-main: ${project.groupId}.${project.artifactId}.${project.name}
-version: ${loritime.version}
-api-version: ${minecraft.api.version}
-description: ${project.description}
-website: 'https://github.com/Lorias-Jak/LoriTime'
-authors: [Lorias-Jak]
