+++ paper/.editorconfig
