+++ paper/.gitattributes
