+++ paper/.gitignore
